requirements.txt
absl-py==2.1.0
accelerate==0.23.0
aiobotocore==2.5.4
aiohttp==3.9.1
aioitertools==0.11.0
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
black @ git+https://github.com/psf/black.git@d1d4fc58d3b744db35ac832cddf1d076ba5f5e7f
bleach==6.0.0
blinker==1.6.3
bokeh==3.3.0
botocore==1.31.17
bs4==0.0.1
cachetools==5.3.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cloudpickle==3.0.0
comm==0.1.4
contourpy==1.1.0
cycler==0.11.0
dask==2023.10.0
debugpy==1.6.7.post1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
distributed==2023.10.0
docopt==0.6.2
editdistance==0.6.2
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.18.0
filelock==3.13.1
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
flatbuffers==23.5.26
fonttools==4.42.1
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.9.2
gast==0.5.4
google-auth==2.28.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
greenlet==3.0.3
grip==4.6.2
grpcio==1.60.1
h5py==3.10.0
huggingface-hub==0.20.3
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
ipykernel==6.25.1
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.1.0
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.0
jmespath==1.0.1
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonrpcclient==4.0.3
jsonrpcserver==5.0.9
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-ascending==0.1.27
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==7.4.9
jupyter_core==5.3.1
jupyter_server==2.7.2
jupyter_server_terminals==0.4.4
jupyterlab==4.0.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.24.0
jupytext==1.15.2
keras==2.15.0
kiwisolver==1.4.5
libclang==16.0.6
locket==1.0.0
loguru==0.7.2
lxml==4.9.3
lz4==4.3.2
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdit-py-plugins==0.4.0
mdurl==0.1.2
mergekit==0.0.2.1
mistune==3.0.2
ml-dtypes==0.2.0
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
mypy-extensions==1.0.0
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.16.1
nbformat==5.9.2
nest-asyncio==1.5.7
networkx==3.2.1
notebook==6.5.6
notebook_shim==0.2.3
numpy==1.25.2
oauthlib==3.2.2
opt-einsum==3.3.0
OSlash==0.6.3
overrides==7.4.0
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
partd==1.4.1
path-and-address==2.0.1
pathspec==0.12.1
patsy==0.5.4
peft==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
plotly==5.16.1
pluggy==1.4.0
polars==0.19.12
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.25.2
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pybind11==2.11.1
pycparser==2.21
pydantic==2.4.0
pydantic_core==2.10.0
Pygments==2.16.1
pynvim==0.5.0
pyparsing==3.0.9
pytest==8.0.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3
pyxet==0.1.8
PyYAML==6.0.1
pyzmq==24.0.1
qtconsole==5.4.3
QtPy==2.3.1
referencing==0.30.2
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.9.2
rsa==4.9
s3fs==2023.9.2
safetensors==0.3.3
scikit-learn==1.3.2
SciPy==1.11.1
seaborn==0.13.1
Send2Trash==1.8.2
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==2.0.22
stack-data==0.6.2
statsmodels==0.14.0
sympy==1.12
tabulate==0.9.0
tblib==3.0.0
tenacity==8.2.3
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.36.0
tensorflow-macos==2.15.0
termcolor==2.4.0
terminado==0.17.1
testpath==0.6.0
threadpoolctl==3.2.0
tinycss2==1.2.1
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.0
torch==2.2.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
transformers==4.37.1
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.18
virtualenv==20.25.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.2
Werkzeug==3.0.1
widgetsnbextension==4.0.8
wordcloud==1.9.3
wrapt==1.14.1
xyzservices==2023.10.0
yarl==1.9.3
zict==3.0.0
zipp==3.17.0
